% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datsu_api.R
\name{datsu_api}
\alias{datsu_api}
\title{Build a DATSU web service url}
\usage{
datsu_api(service, ...)
}
\arguments{
\item{service}{the name of the service}

\item{...}{name arguments will be added as queries}
}
\value{
a complete url as a character string
}
\description{
utility to build a url with optional query arguments
}
\examples{

datsu_api("hi", bye = 21)
datsu_api("getDataverIDs")

}
