% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataFieldsDescription.R
\name{getDataFieldsDescription}
\alias{getDataFieldsDescription}
\title{get list of datasets}
\usage{
getDataFieldsDescription(datasetverID, recordType = NULL)
}
\arguments{
\item{datasetverID}{the dataset ID}

\item{recordType}{string name of the record type, optional}
}
\value{
List of the supported records and fields.
}
\description{
A list of the available fields for the specified format and record (optional filter).
in this list it is specified the field type, if it is mandatory and if it is linked with any
vocabulary.
}
\examples{

\donttest{
getDataFieldsDescription(145)
getDataFieldsDescription(145, "VE")
}
}
