% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataverIDs.R
\name{getDataverIDs}
\alias{getDataverIDs}
\title{get list of datasets}
\usage{
getDataverIDs()
}
\value{
The list of Datasets that can be screened in DATSU with the IDs
}
\description{
This API allows the user to have a list of all the datasets IDs that can be screened in the DATSU API.
}
\examples{
\donttest{
formats <- getDataverIDs()
formats[grep("vms", tolower(formats$description)), ]
}
}
