% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListQCChecks.R
\name{getListQCChecks}
\alias{getListQCChecks}
\title{get list of datasets}
\usage{
getListQCChecks(datasetverID, recordType = NULL)
}
\arguments{
\item{datasetverID}{the dataset ID}

\item{recordType}{string name of the record type, optional}
}
\value{
The list of Datasets that can be screened in DATSU with the IDs
}
\description{
This API allows the user to have a list of all the datasets IDs that can be screened in the DATSU API.
}
\examples{
\donttest{
getListQCChecks(145)
qc_ve <- getListQCChecks(145, "VE")
}
}
