% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMySessionsList.R
\name{getMySessionsList}
\alias{getMySessionsList}
\title{get screening session list for a user}
\usage{
getMySessionsList()
}
\value{
The details of the users sessions
}
\description{
This API of the web service returns a list of the sessions for the authenticated user.
}
\examples{

\dontrun{
getMySessionsList()
}
}
