% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecordIDs.R
\name{getRecordIDs}
\alias{getRecordIDs}
\title{get list of records for a dataset}
\usage{
getRecordIDs(datasetverID)
}
\arguments{
\item{datasetverID}{the dataset ID}
}
\value{
List of the supported records for the specified records
}
\description{
A list of the available records for the specified format
}
\examples{
\donttest{
getRecordIDs(145)
}
}
