% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScreeningSessionDetails.R
\name{getScreeningSessionDetails}
\alias{getScreeningSessionDetails}
\title{get screening session details}
\usage{
getScreeningSessionDetails(sessionID)
}
\arguments{
\item{sessionID}{the session ID}
}
\value{
Details of the session
}
\description{
This API allows the user to view the details of a specific session.
}
\examples{
\donttest{
getScreeningSessionDetails(10)
}
}
