% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScreeningSessionMessages.R
\name{getScreeningSessionMessages}
\alias{getScreeningSessionMessages}
\title{get screening session messages}
\usage{
getScreeningSessionMessages(sessionID)
}
\arguments{
\item{sessionID}{the session ID}
}
\value{
List of the messages for the specified session
}
\description{
A list of the messages for the specified session ID
}
\examples{
\donttest{
messages <- getScreeningSessionMessages(10)
}
}
