% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScreeningSessionsList.R
\name{getScreeningSessionsList}
\alias{getScreeningSessionsList}
\title{get screening session list}
\usage{
getScreeningSessionsList(dataSetVerID, year = NULL, country = NULL)
}
\arguments{
\item{dataSetVerID}{the dataset version id}

\item{year}{filter by year, optional}

\item{country}{filter by country, optional}
}
\value{
The details of the sessions
}
\description{
This API allows the user to screen a file using the API
}
\examples{
\donttest{
getScreeningSessionsList(145)
getScreeningSessionsList(145, year = 2020)
}
}
