% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadDatsuFileFireAndForget.R
\name{uploadDatsuFileFireAndForget}
\alias{uploadDatsuFileFireAndForget}
\title{get list of records for a dataset}
\usage{
uploadDatsuFileFireAndForget(
  fileToUpload,
  emailAddress,
  dataSetVerID,
  sendEmail = TRUE,
  errorLimit = 30000
)
}
\arguments{
\item{fileToUpload}{(stream accepted as IFormFile, part of the body of the request)}

\item{emailAddress}{email address of the user}

\item{dataSetVerID}{The version of the dataset}

\item{sendEmail}{it is set to TRUE by default, this will specify if the user will to receive an
email when the session is finished}

\item{errorLimit}{does not need to be specified, it is 30000 by default}
}
\value{
The ID of the DATSU session
}
\description{
This API allows the user to screen a file using the API.
}
