% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datsu_api.R
\name{datsu_api}
\alias{datsu_api}
\title{Build the api url for given service}
\usage{
datsu_api(service, ...)
}
\arguments{
\item{service}{the name of the service}

\item{...}{name arguments will be added as queries}
}
\description{
utility to build an api url with optional query arguemnts
}
\examples{

datsu_api("hi", bye = 21)
datsu_api("getDataverIDs")

}
